requirements.txt
# Linux Kernel Optimization Framework - Requirements
# Install with: pip install -r requirements.txt

# Core scientific computing
numpy>=1.20.0
scipy>=1.7.0

# Bayesian Optimization
scikit-optimize>=0.9.0

# Genetic Algorithm
pygad>=3.0.0

# System monitoring
psutil>=5.8.0

# Configuration management
PyYAML>=5.4.0

pathlib>=1.0.0

matplotlib>=3.3.0

requests

dotenv

# Optional dependencies for enhanced functionality
# matplotlib>=3.3.0  # For plotting (optional)
# scikit-learn>=1.0.0  # For additional ML algorithms (optional)
# pandas>=1.3.0  # For data analysis (optional)

# Development and testing (optional)
# pytest>=6.0.0
# flake8>=3.9.0
# black>=21.0.0
